+++ tests/test-dirs/occurrences/issue1404.t
@@ -1,6 +1,3 @@
-FIXME there is a discrepancy on the detection of the expression under the cursor
-between locate and occurrences.
-
 occurrences identifier-at 2:0 returns the occurrences of [x]
   $ $MERLIN single occurrences -identifier-at 2:0 -filename opt.ml <<EOF | \
   > jq '.value'
@@ -63,6 +60,37 @@ occurrences identifier-at 2:1 returns the occurrences of [+]
     }
   ]
 
+occurrences identifier-at 2:2 returns the occurrences of [y]
+  $ $MERLIN single occurrences -identifier-at 2:2 -filename opt.ml <<EOF | \
+  > jq '.value'
+  > let x = 3 and y = 4 + 2 in
+  > x+y
+  > EOF
+  [
+    {
+      "start": {
+        "line": 1,
+        "col": 14
+      },
+      "end": {
+        "line": 1,
+        "col": 15
+      },
+      "stale": false
+    },
+    {
+      "start": {
+        "line": 2,
+        "col": 2
+      },
+      "end": {
+        "line": 2,
+        "col": 3
+      },
+      "stale": false
+    }
+  ]
+
 locate position 2:0 returns the definition of [x]
   $ $MERLIN single locate -position 2:0 -filename opt.ml <<EOF | \
   > jq '.value'
@@ -89,5 +117,19 @@ locate position 2:1 returns the definition of [(+)]
     "pos": {
       "line": 304,
       "col": 9
+    }
+  }
+
+locate position 2:2 returns the definition of [y]
+  $ $MERLIN single locate -position 2:2 -filename opt.ml <<EOF | \
+  > jq '.value'
+  > let x = 3 and y = 4 + 2 in
+  > x+y
+  > EOF
+  {
+    "file": "opt.ml",
+    "pos": {
+      "line": 1,
+      "col": 14
     }
   }